+++ lib/parsers/courses.rb
@@ -1,1 +1,5 @@
+module Parsers
+  module Courses
 
+  end
+end
