+++ lib/parsers/projects.rb
@@ -1,1 +1,5 @@
+module Parsers
+  module Projects
 
+  end
+end
